Update Account and Book windows.
Introduce AccountInfo and BookInfo classes that store the
state of the inputs from the user to separate the access
between UI elements and the actual information entered by the user.

The change will help with changes in the future to validate the
user input prior to inserting records to the database.

Github issues: #3 and #4